+++ NEWS.md
@@ -1,3 +1,12 @@
+1.19.1 - 2014-12-15
+-------------------
+
+Bug fixes:
+
+ * Closes stream responses on error (#188)
+ * Closes fd for log files on `kill -HUP` (#187)
+
+
 1.19.0 - 2014-11-21
 -------------------